Marty.renderToString - Re-render until no new fetches
In Marty.renderToString we

1. Render a component tree
2. Check to see if any components made fetches and wait for them
3. Re-render the component tree

This makes the assumption that your components will not make subsequent
fetches as the result of the initial fetch. This assumption falls down
because you might have a child component that needs to fetch state but
it is only rendered once the parent component has finished fetching its
state.

Marty.renderToString will re-render the component until it no longer
creates new fetches. To avoid an infinite loop we will only re-render 10
times however you can configure this by passing in `{
maxNumberOfIterations: 100 }`